fix: move LiveCursorContainer to its own file and don't refactor a Point interface since the liveblocks.config.ts types can't have nested types
